app/requirements.txt
cython==0.28.6
pygame
kivy
sphinx
buildozer
git+http://github.com/HeaTTheatR/KivyMD@08692d7b51927ffab7124972c6214fb36492bc85#egg=kivymd
docutils
pygments
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.32.1
googleapis-common-protos==1.5.9
grpcio==1.19.0
idna==2.8
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1